pyproject.toml
[tool.poetry]
name = "metron-sites"
version = "0.1.0"
description = ""
authors = ["Quam Kponton <quam.kponton@gmail.com>"]
readme = "README.md"
packages = [{include = "metron_sites"}]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.2.2"
Flask-Caching = "^2.0.1"
Flask-Cors = "^3.0.10"
Flask-Script = "^2.0.6"
MarkupSafe = "^2.1.1"
cryptography = "^38.0.3"
flask-sqlalchemy = "^3.0.2"
connexion = {extras = ["swagger-ui"], version = "^2.14.1"}
psycopg2 = "^2.9.5"


[tool.poetry.group.dev.dependencies]
pytest = "^7.1.3"
coverage = "^6.5.0"
flake8 = "^5.0.4"
pytest-django = "^4.5.2"
mypy = "^0.982"
isort = "^5.10.1"
pep8-naming = "^0.13.2"
pylint = "^2.15.3"
pytest-mock = "^3.10.0"
Flask-Testing = "^0.8.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.semantic_release]
version_variable = [
    "pyproject.toml:version",
    "tox.ini:current_version"
]
branch = "main"
upload_to_pypi = true
upload_to_release = true
build_command = "poetry build"
hvcs_domain = "https://github.com/qkponton/"
hvcs_api_domain = "https://github.com/qkponton/"
repository = "metron-sites"
